+++ src/main/java/org/zalando/nakadi/filters/LoggingFilter.java
@@ -1,16 +1,23 @@
 package org.zalando.nakadi.filters;
 
+import com.google.common.base.Charsets;
 import com.google.common.net.HttpHeaders;
+import org.apache.commons.codec.binary.Hex;
+import org.json.JSONObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
+import org.zalando.nakadi.service.NakadiKpiPublisher;
 
 import javax.servlet.FilterChain;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.security.MessageDigest;
 import java.security.Principal;
 import java.util.Optional;
 
@@ -18,6 +25,18 @@ import java.util.Optional;
 public class LoggingFilter extends OncePerRequestFilter {
 
     private static final Logger LOG = LoggerFactory.getLogger(LoggingFilter.class);
+    private final MessageDigest sha256MessageDigest;
+    private final NakadiKpiPublisher nakadiKpiPublisher;
+    private final String accessLogEventType;
+
+    @Autowired
+    public LoggingFilter(final MessageDigest sha256MessageDigest,
+                         final NakadiKpiPublisher nakadiKpiPublisher,
+                         @Value("${nakadi.kpi.event-types.nakadiAccessLog}") final String accessLogEventType) {
+        this.sha256MessageDigest = sha256MessageDigest;
+        this.nakadiKpiPublisher = nakadiKpiPublisher;
+        this.accessLogEventType = accessLogEventType;
+    }
 
     @Override
     protected void doFilterInternal(final HttpServletRequest request,
@@ -50,6 +69,14 @@ public class LoggingFilter extends OncePerRequestFilter {
                     timing,
                     contentEncoding,
                     acceptEncoding);
+            nakadiKpiPublisher.publish(accessLogEventType, () -> new JSONObject()
+                    .put("method", method)
+                    .put("path", path)
+                    .put("query", query)
+                    .put("app", user)
+                    .put("app_hashed", Hex.encodeHexString(sha256MessageDigest.digest(user.getBytes(Charsets.UTF_8))))
+                    .put("status_code", response.getStatus())
+                    .put("response_time_ms", timing));
         }
     }
 }