+++ src/main/java/org/zalando/nakadi/service/subscription/state/State.java
@@ -2,11 +2,13 @@ package org.zalando.nakadi.service.subscription.state;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.zalando.nakadi.domain.NakadiCursor;
 import org.zalando.nakadi.service.subscription.StreamParameters;
 import org.zalando.nakadi.service.subscription.StreamingContext;
 import org.zalando.nakadi.service.subscription.SubscriptionOutput;
 import org.zalando.nakadi.service.subscription.zk.ZkSubscriptionClient;
 
+import java.util.Comparator;
 import java.util.concurrent.TimeUnit;
 
 public abstract class State {
@@ -82,5 +84,9 @@ public abstract class State {
 
     public StreamingContext getContext() {
         return context;
+    }
+
+    public Comparator<NakadiCursor> getComparator() {
+        return getContext().getCursorComparator();
     }
 }