+++ src/main/java/org/zalando/nakadi/service/timeline/TimelineService.java
@@ -10,6 +10,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.TransactionException;
 import org.springframework.transaction.support.TransactionTemplate;
 import org.zalando.nakadi.config.NakadiSettings;
+import org.zalando.nakadi.domain.DefaultStorage;
 import org.zalando.nakadi.domain.EventType;
 import org.zalando.nakadi.domain.EventTypeBase;
 import org.zalando.nakadi.domain.EventTypeResource;
@@ -43,6 +44,7 @@ import org.zalando.nakadi.repository.db.EventTypeCache;
 import org.zalando.nakadi.repository.db.StorageDbRepository;
 import org.zalando.nakadi.repository.db.TimelineDbRepository;
 import org.zalando.nakadi.service.AdminService;
+import org.zalando.nakadi.service.NakadiCursorComparator;
 
 import javax.annotation.Nullable;
 import java.util.Collections;
@@ -63,7 +65,7 @@ public class TimelineService {
     private final TimelineDbRepository timelineDbRepository;
     private final TopicRepositoryHolder topicRepositoryHolder;
     private final TransactionTemplate transactionTemplate;
-    private final Storage defaultStorage;
+    private final DefaultStorage defaultStorage;
     private final AdminService adminService;
 
     @Autowired
@@ -74,7 +76,7 @@ public class TimelineService {
                            final TimelineDbRepository timelineDbRepository,
                            final TopicRepositoryHolder topicRepositoryHolder,
                            final TransactionTemplate transactionTemplate,
-                           @Qualifier("default_storage") final Storage defaultStorage,
+                           @Qualifier("default_storage") final DefaultStorage defaultStorage,
                            final AdminService adminService) {
         this.eventTypeCache = eventTypeCache;
         this.storageDbRepository = storageDbRepository;
@@ -128,11 +130,12 @@ public class TimelineService {
             RepositoryProblemException,
             DuplicatedTimelineException,
             TimelineException {
-        final TopicRepository repository = topicRepositoryHolder.getTopicRepository(defaultStorage);
+        final TopicRepository repository = topicRepositoryHolder.getTopicRepository(defaultStorage.getStorage());
         final String topic = repository.createTopic(partitionsCount, retentionTime);
 
         try {
-            final Timeline timeline = Timeline.createTimeline(eventTypeName, 1, defaultStorage, topic, new Date());
+            final Timeline timeline = Timeline.createTimeline(eventTypeName, 1,
+                    defaultStorage.getStorage(), topic, new Date());
             timeline.setSwitchedAt(new Date());
             timelineDbRepository.createTimeline(timeline);
             eventTypeCache.updated(eventTypeName);
@@ -207,13 +210,15 @@ public class TimelineService {
 
     public EventConsumer createEventConsumer(@Nullable final String clientId, final List<NakadiCursor> positions)
             throws NakadiException, InvalidCursorException {
-        final MultiTimelineEventConsumer result = new MultiTimelineEventConsumer(clientId, this, timelineSync);
+        final MultiTimelineEventConsumer result = new MultiTimelineEventConsumer(
+                clientId, this, timelineSync, new NakadiCursorComparator(eventTypeCache));
         result.reassign(positions);
         return result;
     }
 
     public EventConsumer.ReassignableEventConsumer createEventConsumer(@Nullable final String clientId) {
-        return new MultiTimelineEventConsumer(clientId, this, timelineSync);
+        return new MultiTimelineEventConsumer(
+                clientId, this, timelineSync, new NakadiCursorComparator(eventTypeCache));
     }
 
     private void switchTimelines(final Timeline activeTimeline, final Timeline nextTimeline)
@@ -269,7 +274,7 @@ public class TimelineService {
             NoSuchEventTypeException {
         LOG.info("Deleting all timelines for event type {}", eventTypeName);
         final Multimap<TopicRepository, String> topicsToDelete = ArrayListMultimap.create();
-        for (final Timeline timeline : getActiveTimelinesOrdered(eventTypeName)) {
+        for (final Timeline timeline : getAllTimelinesOrdered(eventTypeName)) {
             topicsToDelete.put(getTopicRepository(timeline), timeline.getTopic());
             timelineDbRepository.deleteTimeline(timeline.getId());
         }