+++ src/main/java/org/zalando/nakadi/util/FeatureToggleService.java
@@ -36,7 +36,9 @@ public interface FeatureToggleService {
         CHECK_PARTITIONS_KEYS("check_partitions_keys"),
         CHECK_OWNING_APPLICATION("check_owning_application"),
         LIMIT_CONSUMERS_NUMBER("limit_consumers_number"),
-        SEND_BATCH_VIA_OUTPUT_STREAM("send_batch_via_output_stream");
+        SEND_BATCH_VIA_OUTPUT_STREAM("send_batch_via_output_stream"),
+        REMOTE_TOKENINFO("remote_tokeninfo"),
+        KPI_COLLECTION("kpi_collection");
 
         private final String id;